+++ medic-doctor/src/main.rs
@@ -4,7 +4,7 @@ use medic_doctor::cli::CliArgs;
 use medic_doctor::run_checks;
 use medic_src::config::Manifest;
 use medic_src::context::Context;
-use medic_src::AppResult;
+use medic_src::{theme, AppResult};
 
 use clap::{CommandFactory, Parser};
 use clap_complete::generate;
@@ -15,6 +15,7 @@ use std::panic;
 fn main() -> AppResult<()> {
   let context = Context::new();
   let cli_args = CliArgs::parse();
+  theme::set_theme((&cli_args.theme).into());
 
   if let Some(completion) = cli_args.completion {
     let mut cmd = CliArgs::command();