+++ medic-outdated/src/main.rs
@@ -2,7 +2,7 @@ use medic_outdated::cli::CliArgs;
 use medic_outdated::run_checks;
 use medic_src::config::Manifest;
 use medic_src::context::Context;
-use medic_src::AppResult;
+use medic_src::{theme, AppResult};
 
 use clap::{CommandFactory, Parser};
 use clap_complete::generate;
@@ -13,6 +13,7 @@ use std::panic;
 fn main() -> AppResult<()> {
   let context = Context::new();
   let cli_args = CliArgs::parse();
+  theme::set_theme((&cli_args.theme).into());
 
   if let Some(completion) = cli_args.completion {
     let mut cmd = CliArgs::command();