+++ medic-run/src/cli/app.rs
@@ -1,5 +1,6 @@
 use clap::Parser;
 use clap_complete::Shell;
+use medic_src::theme::Theme;
 
 #[derive(Debug, Parser)]
 #[clap(author, version, about)]
@@ -19,4 +20,8 @@ pub struct CliArgs {
   /// Shell to generate completions for
   #[clap(long, value_enum, value_parser)]
   pub completion: Option<Shell>,
+
+  /// Color theme
+  #[arg(short, long, env = "MEDIC_THEME", default_value = "auto")]
+  pub theme: Theme,
 }